+++ app/Filament/Resources/Users/Pages/ListUsers.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace App\Filament\Resources\UserResource\Pages;
+namespace App\Filament\Resources\Users\Pages;
 
 use Filament\Actions\CreateAction;
-use App\Filament\Resources\UserResource;
-use Filament\Actions;
+use App\Filament\Resources\Users\UserResource;
 use Filament\Resources\Pages\ListRecords;
 
 class ListUsers extends ListRecords
