+++ app/Http/Controllers/Api/V1/LatestResult.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Http\Resources\V1\ResultResource;
+use App\Http\Resources\V1\Results;
 use App\Models\Result;
 use Http\Discovery\Exception\NotFoundException;
 use Illuminate\Http\Request;
@@ -46,7 +46,7 @@ class LatestResult extends ApiController
             });
 
         return self::sendResponse(
-            data: new ResultResource($result),
+            data: new Results($result),
         );
     }
 }
